TOOLS/Setup: Inform the user about existing setups
If the user tries to create a new setup with the same name of an
existing ones an error message will be generated and it will be
blocked.

+++ tools/setupfiletool/treewidget.py
@@ -280,6 +280,11 @@ class TreeWidget(TreeWidgetContextMenu):
                                     dlg.currentInstrument(),
                                     'setups',
                                     fileName)
+            if abspath in [i.abspath for i in
+                           setupcontroller.setup_directories[
+                               dlg.currentInstrument()]]:
+                QMessageBox.warning(self, 'Error', 'Setup already exists!')
+                return None, None
             try:
                 open(abspath, 'w').close()
             except IOError:
